+++ src/Persistence/Schema/MigrationList.php
@@ -18,6 +18,7 @@ use EnergyCRM\Persistence\Schema\Migrations\AddContractAddresses;
 use EnergyCRM\Persistence\Schema\Migrations\AddContractListIndexes;
 use EnergyCRM\Persistence\Schema\Migrations\AddCustomerAfmIndex;
 use EnergyCRM\Persistence\Schema\Migrations\AddComboEnergyContactColumns;
+use EnergyCRM\Persistence\Schema\Migrations\AddComboMobileContactColumns;
 use EnergyCRM\Persistence\Schema\Migrations\AddCustomerPhoneIndex;
 use EnergyCRM\Persistence\Schema\Migrations\AddForeignKeys;
 use EnergyCRM\Persistence\Schema\Migrations\AddGuaranteeRuleForeignKeys;
@@ -87,6 +88,7 @@ final class MigrationList
             new DedupeProgramsByCode(),
             new AddComboEnergyContactColumns(),
             new AddFileKindSourceColumns(),
+            new AddComboMobileContactColumns(),
         ];
     }
 }